requirements.txt
# PyWebView Desktop App Dependencies

# Core dependencies
pywebview>=5.0.0

# ==================== Platform-specific Dependencies ====================

# Windows recommended (required for better performance)
pywin32>=306; sys_platform == 'win32'

# macOS recommended (optional for better performance)
pyobjc>=10.0; sys_platform == 'darwin'

# Linux recommended (requires GTK installed on system)
# PyGObject>=3.46.0; sys_platform == 'linux'

# ==================== FlowTest Pro Test Platform Dependencies ====================

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0

# Automation
pyautogui>=0.9.54
pygetwindow>=0.0.9

# AI Model Interface
openai>=1.0.0

# Data Processing
numpy>=1.24.0
